Make Terraform config useable
When saving a terraform buffer, 'terraform fmt' will now be called
through an autocmd.
Enforce defaults in terraform LSP config; I suspected they were getting
overwritten, but it seems not to be the case as there is no real
difference with these.
Change the root finder to ignore '.git' since my work is mostly
monorepo-style with several TF roots in a single repo.
Add 'hcl' as a TreeSitter language to keep updated, so that Terraform
highlighting works.