Terraform LSP autocmd fixed for 0.8+
This commit fixes an error complaining that 'command' or 'callback' should not be empty.
The problem was that upgrading to 0.9 caused the formatting_sync() call to be removed, so Lua thought the callback was not set (empty).
This is fixed by using the new name for the format api, .format().
See: https://github.com/hashicorp/terraform-ls/blob/main/docs/USAGE.md#neovim-v080

Also expanded the filetypes to tfvars files as well.